backend/pyproject.toml
[project]
name = "url-shortener"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "Andrew Guest", email = "andrew@agueset.me" }]
dependencies = [
    "litestar>=2.3.2",
    "gunicorn>=21.2.0",
    "uvicorn>=0.24.0.post1",
    "pydantic>=2.5.2",
    "redis>=5.0.1",
    "python-dotenv>=1.0.0",
    "newrelic>=9.2.0",
]
readme = "README.md"
requires-python = ">= 3.11"
license = { text = "MIT" }

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
